requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.0.0
pymongo>=4.0.0
textblob>=0.17.1
streamlit-lottie>=0.0.3
altair>=4.2.0
seaborn>=0.12.2
matplotlib>=3.7.0
groq>=0.1.0
